+++ lib/decorators/custom/Middleware.ts
@@ -1,3 +1,4 @@
+import { Constructor } from "../../interface";
 import { Action } from "../internal/Action";
 import { Injectable } from "../internal/Injectable";
 import { ControllerMetadata } from "./Controller";
@@ -5,17 +6,17 @@ import { RouterMetadata } from "./RouterMethods";
 
 // action middleware
 export function Middleware(middlewares: Function[]) {
-  return function ActionDescriptorDecorator(target: Object, key?: string, descriptor?: PropertyDescriptor) {
-    return key ? ActionMiddleware(middlewares) : RouterGroupMiddleware(middlewares);
+  return (target: any, key?: any, des?: any) => {
+    key ? ActionMiddleware(middlewares, target, key, des) : RouterGroupMiddleware(middlewares, target);
   };
 }
 
-function ActionMiddleware(middlewares: Function[]) {
+function ActionMiddleware(middlewares: Function[], target: Object, key: string, des: PropertyDescriptor) {
   const callback = (target: Object, key: string, des: PropertyDescriptor): Partial<RouterMetadata> => ({ middlewares });
-  return Action(callback);
+  Action(callback)(target, key, des);
 }
 
-function RouterGroupMiddleware(middlewares: Function[]) {
+function RouterGroupMiddleware(middlewares: Function[], target: Constructor) {
   const callback = (): Partial<ControllerMetadata> => ({ middlewares });
-  return Injectable(callback);
+  Injectable(callback)(target);
 }