+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-[2020/09/09] v0.1.7
+[2020/09/10] v0.2.0
+
+- update readme
+- action level middleware
+- controller level middleware
+- change app.useGlobalMiddleware param similar to koa
+
+[2020/09/10] v0.1.7
 
 - separate initializer from ioc
 - update ioc resolve method